operator/testing: simplify cluster suite entrypoints
