+++ .github/workflows/pre-merge.yml
@@ -107,6 +107,8 @@ jobs:
       contents: read
     steps:
     - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd  # v6.0.2
+      with:
+        fetch-depth: 0  # clustertest derives its alpha version from v1.3.0
     - uses: docker/setup-buildx-action@e468171a9de216ec08956ac3ada2f0791b6bd435 #v3.11.1
     - name: Run operator checks
       uses: ./.github/actions/check-deploy-component
