--- deploy/operator/internal/controller/clusterenv_stub_test.go
@@ -1,12 +1,0 @@
-//go:build !clustertest
-
-/*
- * SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package controller
-
-func runClusterTestEnv(runner interface{ Run() int }) int {
-	return runner.Run()
-}
